Basic RSA implementation
I have written an RSA implementation that can generate key pairs,
encrypt and decrypt messages. However, it struggles to generate
sufficiently large key pairs and encrypt or decrypt messages in a
reasonable amount of time - so it is unsuitable for actual use. This is
due to the naive methods used for testing primality and exponentiation.
More efficient algorithms are presented in the RSA paper and are obvious
areas of improvement for the project.